Add `depth` as an additional variable in Sv dataset (#738)
* add first version of swapping dims to freq from channel

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* prototyping function to add depth variable

* fix a few things in function

* fix type

* remove utils/common and move content to consolidate/api

* clean up remnants of moving functions

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* add test and fix bugs

* Update echopype/consolidate/api.py


* Update echopype/consolidate/api.py


* simplify add_depth to use only scalar inputs, udpate test

* add notes in docstring

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci
