+++ echopype/consolidate/__init__.py
@@ -1,3 +1,3 @@
-from .api import add_location
+from .api import add_depth, add_location, swap_dims_channel_frequency
 
-__all__ = ["add_location"]
+__all__ = ["swap_dims_channel_frequency", "add_depth", "add_location"]
