+++ echopype/preprocess/__init__.py
@@ -1,8 +1,7 @@
-from .api import compute_MVBS, compute_MVBS_index_binning, remove_noise, swap_dims_channel_frequency
+from .api import compute_MVBS, compute_MVBS_index_binning, remove_noise
 
 __all__ = [
     "compute_MVBS",
     "compute_MVBS_index_binning",
     "remove_noise",
-    "swap_dims_channel_frequency",
 ]
